+++ R/ssyt.R
@@ -17,10 +17,10 @@ count_ssytx <- function(lambda, n) {
   n <- as.integer(n)
   stopifnot(n >= 1L)
   if(n < length(lambda)) {
-    return(0)
+    return(0L)
   }
   if(n == 1L) {
-    return(1)
+    return(1L)
   }
   lambda <- c(lambda, rep(0, n-length(lambda)))
   out <- 1
@@ -29,11 +29,7 @@ count_ssytx <- function(lambda, n) {
       out <- out * (1 + (lambda[i] - lambda[j]) / (j - i)) 
     }
   }
-  out
-}
-
-rg <- function(start, end) {
-  seq_len(end - start + 1L) + (start - 1L)
+  as.integer(ceiling(out))
 }
 
 #' Enumeration of semistandard Young tableaux
@@ -49,47 +45,35 @@ rg <- function(start, end) {
 #' @export
 #'
 #' @examples
-#' all_ssytx(c(2, 1), 3)
+#' ssytx <- all_ssytx(c(2, 1), 3)
+#' lapply(ssytx, prettyTableau)
 all_ssytx <- function(lambda, n) {
-  lambda <- as.integer(checkPartition(lambda))
-  n <- as.integer(n)
-  stopifnot(n >= 1L)
-  l <- length(lambda)
-  ll <- lambda[l]
-  tmp <- list(list())
-  res <- list()
-  while(length(tmp) > 0L) {
-    T <- tmp[[1L]]
-    tmp <- tmp[-1L]
-    k <- length(T)
-    if(k == l && length(T[[k]]) >= ll) {
-      res <- c(res, list(T))
+  lambda <- checkPartition(lambda)
+  row <- function(n, len, prev, xxs) {
+    if(len == 0L) {
+      list(integer(0L))
     } else {
-      if(k == 0L || length(T[[k]]) %in% c(0L, lambda[k])) {
-        if(k == 0L) {
-          start <- 1L
-        } else {
-          start <- T[[k]][1L] + 1L
-        }
-        for(u in rg(start, n)) {
-          U <- c(T, list(u))
-          tmp <- c(tmp, list(U))
-        }
-      } else {
-        Tk <- T[[k]]
-        lk <- length(Tk)
-        if(k == 1L) {
-          start <- Tk[lk]
-        } else {
-          start <- max(Tk[lk], T[[k-1L]][lk] + 1L)
-        }
-        for(u in rg(start, n)) {
-          U <- T
-          U[[k]] <- c(Tk, u)
-          tmp <- c(tmp, list(U))
-        }
-      }
+      x <- xxs[[1L]]
+      xs <- xxs[-1L]
+      do.call(c, lapply(.rg(max(x, prev), n), function(a) {
+        lapply(row(n, len - 1L, a, xs), function(as) {
+          c(a, as)
+        })
+      }))
     }
   }
-  res
-} 
+  worker <- function(prevRow, sss) {
+    if(length(sss) == 0L) {
+      list(list())
+    } else {
+      s <- sss[[1L]]
+      ss <- sss[-1L]
+      do.call(c, lapply(row(n, s, 1L, prevRow), function(r) {
+        lapply(worker(r + 1L, ss), function(rs) {
+          c(list(r), rs)
+        })
+      }))
+    }
+  }
+  worker(rep(0L, lambda[1L]), lambda)
+}