+++ NEWS.md
@@ -1,3 +1,14 @@
+# syt 0.4.0
+
+* New functions for general tableaux.
+
+* New functions for skew tableaux.
+
+* Kostka numbers.
+
+* There was a bug in the function `all_ssytx`.
+
+
 # syt 0.3.0
 
 * Semistandard Young tableaux: enumeration and count.
@@ -6,12 +17,10 @@
 # syt 0.2.0
 
 * Fixed a small mistake.
+
 * Improved the efficiency of the enumeration of all standard Young tableaux.
 
 
 # syt 0.1.0
 
 * First release.
-
-
-